feat: create the spacemacs-literate-layering package file
In this commit, we add the main file of the package,
=spacemacs-literate-layering.el=. The new functions are simply slight
modifications of the =configuration-layer//create-layer= and
=configuration-layer//copy-template= functions from Spacemacs. The main
difference is that instead of copying the individual template files for
~packages.el~ and ~layers.el~, a template Org file will be copied to the layer
directory instead.

The template Org file is what puts the `literate' in the name of the title, as
the layer configuration will be done literally in it and tangled to furnish the
different files of the layer.

The default behaviour mimics that of =configuration-layer//copy-template=. )for
example, where unless =user-full-name= is specified, the default author and
email address are "Sylvain Benner & Contributors" and
"sylvain.benner@gmail.com", respectively.