fix: ensure Org headers are reloaded after copying template
There was an issue where tangling newly generated layer README files does not
create tangled files that furnish the layer names at all, they were left as
%LAYER_NAME%, unsubstituted. This comes from the fact that the source
blocks inherited the layer name from the file property headers when it was first
opened, which was %LAYER_NAME% as it had yet to be substituted.

Fixing this means that once the keyword substitutions were complete, the file
needs to be reloaded for the source blocks to refresh. This commit furnishes the
solution to this issue.