fix: handle cases when buffers are not files
Triggering the minor mode will first check if the buffers are real files before
activating. This is not a hack as Org buffers for literate layer programming are
expected to be real files on the disk, not temporary buffer files. The output of
the try-catch is an empty string as it appears from the debug statement that
Emacs expects a string as the output and fails when it is simply `nil' for some
reason.

Fixes #1